requirements.txt
snowflake-snowpark-python
pandas
numpy
notebook
jinja2
jinjasql
iplantuml

tensorflow
scikit-learn
opencv-python
